Refactored some code out of AddErrorSuppressionModal and into esFunctions so testing is easier
